Web/requirements.txt
robotframework==3.1.2
robotframework-debuglibrary==1.1.4
robotframework-faker==4.2.0
robotframework-seleniumlibrary==4.1.0
robotframework-pageobjectlibrary==1.0.2
Faker==2.0.4
